+++ packages/react/src/components/tree-view/tree-view-root.tsx
@@ -20,22 +20,27 @@ export interface TreeViewRootProps<T extends TreeNode> extends HTMLProps<'div'>,
 const TreeViewImpl = <T extends TreeNode>(props: TreeViewRootProps<T>, ref: React.Ref<HTMLDivElement>) => {
   const [renderStrategyProps, treeViewProps] = splitRenderStrategyProps(props)
   const [useTreeViewProps, localProps] = createSplitProps<UseTreeViewProps<T>>()(treeViewProps, [
+    'checkedValue',
     'collection',
+    'defaultCheckedValue',
     'defaultExpandedValue',
+    'defaultFocusedValue',
     'defaultSelectedValue',
     'expandedValue',
     'expandOnClick',
     'focusedValue',
     'id',
     'ids',
+    'loadChildren',
+    'onCheckedChange',
     'onExpandedChange',
     'onFocusChange',
+    'onLoadChildrenComplete',
+    'onLoadChildrenError',
     'onSelectionChange',
     'selectedValue',
     'selectionMode',
     'typeahead',
-    'loadChildren',
-    'onLoadChildrenComplete',
   ])
 
   const treeView = useTreeView(useTreeViewProps)